feat(cloudflare)!: Remove wrapRequestHandlerWithInit as export
